Add ability to edit retrospective assessment condition
ASRU should be able to add an RA condition if one does not apply, and should see if one has been automatically applied, but it should not appear under the conditions section when viewing the granted licence because RA is already exposed in the "Project summary" section.
